refactor(wework): isolate native browser capture
